evachargese: fix autoboot and reduce timeout
At the moment CONFIG_AUTOBOOT_KEYED is declared in board config file,
but without argument. This prevents that this option is detected as
enabled and thus the feature is not active at all.
Also enable Ctrl+C feature to be in sync with Tarragon boards.

While at, reduce the timeout from 120s to 10s - during normal field
operation, we want to start the board fast, even in case of any
interference on the Debug UART. On development systems, this value
can be manually increased via U-Boot environmenet.

Signed-off-by: Michael Heimpold <michael.heimpold@in-tech.com>