tarragon: enable eMMCs reset pin while booting
This "hack" burns the eMMC's OTP register to enable the reset pin
functionality. It is done before passing control to Linux kernel
so that the Linux kernel's eMMC driver can already make use of
this function. It allows the kernel to directly switch to 1.8V,
then re-animating the eMMC with the reset line and finally to
switch to HS200 mode - which is really required during production
process to optimize speed.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>