thermal: imx: Introduce cooldown hysteresis
The calculation of critical temperature and its behavior is identical
between U-Boot and Linux imx thermal ( critical = maxC - 5C ). This is
problematic because of the following reasons:
* U-Boot usually runs with a lower clock frequency and less parallel tasks
  in comparision to a Linux boot
* thermal management is not available at the very begin of Linux boot
* A/B boot don't expect thermal shutdown during Linux boot, which is
  considered as a failure in the boot image

Because of this, U-Boot must ensure that there is a temperature reserve
to finish the Linux boot without triggering a thermal shutdown. Otherwise
this could lead to some kind of swinging between U-Boot and Linux or even
worse a real CPU shutdown. An additional hysteresis of 10 degress is
considered enough.

Signed-off-by: Stefan Wahren <stefan.wahren@chargebyte.com>